+++ .env.example
@@ -17,7 +17,7 @@ GOOGLE_CLIENT_SECRET=
 SESSION_SECRET=
 
 # Frontend base URL
-CLIENT_URL=http://localhost
+VITE_CLIENT_URL=http://localhost
 
 # Callback URL for OAuth
 CALLBACK_URL=http://localhost/api/auth/google/callback
