+++ frontend/src/types/leaderboard.ts
@@ -1,0 +1,10 @@
+export type Leaderboard = {
+	id: string;
+	name: string;
+	email: string;
+	cfHandle: string;
+	cfRating: number;
+	pfpUrl: string;
+	batch: string;
+	rank?: number;
+};
